删除bug

+++ _1kkk/pipelines.py
@@ -227,9 +227,12 @@ class downloadImage(threading.Thread):
         #路径
         epubpath="./tmp/image/%s/%s"%(manga.id,ci.id)
         #删除冗余数据
-        os.remove("%s.epub"%epubpath)
-        os.remove("%s.mobi"%epubpath)
-        os.remove("%s.zip"%epubpath)
+        if os.path.exists("%s.epub"%epubpath):
+            os.remove("%s.epub"%epubpath)
+        if os.path.exists("%s.mobi"%epubpath):
+            os.remove("%s.mobi"%epubpath)
+        if os.path.exists("%s.zip"%epubpath):
+            os.remove("%s.zip"%epubpath)
         title="%s[%s][%s][%s]"%(manga.name,ci.chid,manga.author,manga.type)
         createKVBook(path,"%s.epub"%epubpath,title)
         #删除epub
